fix open post with incompleteRewards

+++ core/domain/src/main/kotlin/su/afk/kemonos/domain/models/ProfilePosts.kt
@@ -87,7 +87,7 @@ data class VideoDomain(
 
 @Serializable
 data class IncompleteRewards(
-    val price: Int?,
+    val price: Double?,
     val mediaCount: Int?,
     val photoCount: Int?,
     val videoCount: Int?,
